Sépare la création et le démarrage du serveur
Cette séparation va faciliter l'écriture de tests dont on va avoir besoin pour se débarrasser de Nest.
